install mistake: needed to be .ts not .js project
